Added fragments
